# services/openapi.yaml
openapi: 3.0.3
info:
  title: SME Cloud-Native Microservices API - Topic 10
  version: 1.0.0
  description: API contract for gateway routing, authorization enforcement, security testing, and API fuzzing.
servers:
  - url: https://localhost:8443
    description: Kong API Gateway HTTPS endpoint
  - url: http://localhost:8000
    description: Kong API Gateway HTTP endpoint for local testing
...
paths:
...
  /api/v1/orders/{order_id}/vulnerable:
    get:
      tags:
        - Orders
      operationId: getOrderByIdVulnerable
      summary: Get order detail through intentionally vulnerable BOLA endpoint
      description: This endpoint intentionally omits ownership enforcement to demonstrate Broken Object Level Authorization in a controlled lab scenario. It must not be treated as the protected production endpoint.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          description: Order identifier
          schema:
            type: string
          example: "ord-bob-2001"
      responses:
        '200':
          description: Order detail is returned for BOLA demonstration
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/orders/{order_id}/fixed:
    get:
      tags:
        - Orders
      operationId: getOrderByIdFixed
      summary: Get order detail through ownership-protected endpoint
      description: This endpoint enforces server-side ownership or admin authorization and is the mitigation endpoint for the BOLA scenario.
      security:
        - bearerAuth: []
      parameters:
        - name: order_id
          in: path
          required: true
          description: Order identifier
          schema:
            type: string
          example: "ord-alice-1001"
      responses:
        '200':
          description: Order detail is returned when the caller is authorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetail'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/orders/internal/verify-ownership:
    post:
      tags:
        - Orders
      operationId: verifyOrderOwnershipInternal
      summary: Verify order ownership for internal Billing service calls
      description: Internal service-to-service endpoint. Requires a valid Client Credentials access token for `billing-service-client` with role `order-ownership-read`. Human user tokens and other service clients are rejected.
      security:
        - bearerAuth: []
      parameters:
        - name: X-Correlation-ID
          in: header
          required: false
          description: Correlation identifier propagated from Billing to Order.
          schema:
            type: string
          example: "req-123456"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OwnershipVerificationRequest'
      responses:
        '200':
          description: Ownership verification completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OwnershipVerificationResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    Unauthorized:
      description: Missing, expired, revoked, or invalid JWT
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Authenticated principal does not have permission to access this resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Requested resource does not exist
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request body failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    OrderDetail:
      type: object
      required:
        - order_id
        - owner_id
        - amount
        - status
      properties:
        order_id:
          type: string
          example: "order-001"
        owner_id:
          type: string
          example: "alice"
        amount:
          type: number
          format: float
          example: 150000
        status:
          type: string
          example: "created"
...
    OwnershipVerificationRequest:
      type: object
      required:
        - order_id
        - subject
      properties:
        order_id:
          type: string
          example: "ord-alice-1001"
        subject:
          type: string
          example: "alice"
    OwnershipVerificationResponse:
      type: object
      required:
        - order_id
        - allowed
      properties:
        order_id:
          type: string
          example: "ord-alice-1001"
        allowed:
          type: boolean
          example: true
        correlation_id:
          type: string
          example: "req-123456"
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "forbidden"
        message:
          type: string
          example: "Access denied"
        correlation_id:
          type: string
          example: "req-123456"